Rename setupIndex to setup
