+++ sdk/search/search-documents/test/node/searchClient.spec.ts
@@ -15,7 +15,7 @@ import {
   IndexDocumentsBatch
 } from "../../src/index";
 import { Hotel } from "../utils/interfaces";
-import { createIndex, populateIndex, WAIT_TIME } from "../utils/setupIndex";
+import { createIndex, populateIndex, WAIT_TIME } from "../utils/setup";
 import { delay } from "@azure/core-http";
 
 const TEST_INDEX_NAME = "hotel-live-test1";
