+++ sdk/search/search-documents/test/utils/setup.ts
